requirements.txt
appdirs==1.4.3
backports.shutil-get-terminal-size==1.0.0
click==6.7
decorator==4.0.11
enum34==1.1.6
Flask==0.12
Flask-MySQL==1.4.0
Flask-MySQLdb==0.2.0
Flask-OAuth==0.12
Flask-Script==2.0.5
httplib2==0.10.3
ipython==5.3.0
ipython-genutils==0.2.0
itsdangerous==0.24
Jinja2==2.9.5
MarkupSafe==0.23
mysqlclient==1.3.10
oauth2==1.9.0.post1
packaging==16.8
pathlib2==2.2.1
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.14
ptyprocess==0.5.1
Pygments==2.2.0
PyMySQL==0.7.10
pyparsing==2.2.0
requests==2.14.1
scandir==1.5
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.2
wcwidth==0.1.7
Werkzeug==0.11.15
